# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Template API for Golang project",
        "title": "Base APP Go",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/roles/role-access/{role_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves all menu permissions associated with a specific role. Requires Bearer token.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role Menu Permission"
                ],
                "summary": "Get Role Menu Permission",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Role ID",
                        "name": "role_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully get role menu permission",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.PaginatePayload"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_role_presenter.RoleAccessResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid Role ID",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "404": {
                        "description": "Role ID not found",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "500": {
                        "description": "Failed get role menu permission",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates menu permission assigned to a specific role. Requires Bearer token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role Menu Permission"
                ],
                "summary": "Update Role Menu Permission",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Role ID",
                        "name": "role_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Role menu permission payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_role_presenter.UpdateRoleAccessRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully update role menu permission",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.DataPayload"
                        }
                    },
                    "400": {
                        "description": "Bad Request (Invalid ID, Invalid JSON, or Validation Error)",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "404": {
                        "description": "Role ID or Menu Permission not found",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    }
                }
            }
        },
        "/roles/{role_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Fetches the details of a single role by its ID. Requires Bearer token.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "Get Role Detail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Role ID",
                        "name": "role_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully get role detail",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.DataPayload"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_role_presenter.ReadRoleResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request (Invalid Role ID)",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "404": {
                        "description": "Not Found (Role ID not found)",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "500": {
                        "description": "Failed get role detail",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates an existing role by its ID. Requires Bearer token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "Update Role",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Role ID",
                        "name": "role_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Role data to update",
                        "name": "role",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_role_presenter.UpdateRoleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully update role",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.DataPayload"
                        }
                    },
                    "400": {
                        "description": "Bad Request (Invalid ID, Invalid JSON, or Validation Error)",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "404": {
                        "description": "Role ID not found",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "500": {
                        "description": "Failed update role",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a role by its ID. Requires Bearer token.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "Delete Role",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Role ID",
                        "name": "role_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully delete role",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.DataPayload"
                        }
                    },
                    "400": {
                        "description": "Bad Request (Invalid Role ID)",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "401": {
                        "description": "Unauthorized (Missing Auth)",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "404": {
                        "description": "Role ID not found",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "500": {
                        "description": "Failed delete role",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "be-dashboard-nba_internal_presentation_response.DataPayload": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "be-dashboard-nba_internal_presentation_response.ErrorPayload": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "be-dashboard-nba_internal_presentation_response.PaginatePayload": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "paginate": {
                    "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.paginatePayload"
                }
            }
        },
        "be-dashboard-nba_internal_presentation_response.paginatePayload": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total_data": {
                    "type": "integer"
                },
                "total_page": {
                    "type": "number"
                }
            }
        },
        "be-dashboard-nba_internal_presentation_role_presenter.Access": {
            "type": "object",
            "properties": {
                "access_id": {
                    "type": "integer"
                },
                "access_name": {
                    "type": "string"
                },
                "has_access": {
                    "type": "boolean"
                }
            }
        },
...
        "be-dashboard-nba_internal_presentation_role_presenter.ReadRoleResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "be-dashboard-nba_internal_presentation_role_presenter.RoleAccessResponse": {
            "type": "object",
            "properties": {
                "accesses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/be-dashboard-nba_internal_presentation_role_presenter.Access"
                    }
                },
                "menu_id": {
                    "type": "integer"
                },
                "menu_name": {
                    "type": "string"
                }
            }
        },
        "be-dashboard-nba_internal_presentation_role_presenter.UpdateRoleAccessItem": {
            "type": "object",
            "required": [
                "access_id",
                "has_access"
            ],
            "properties": {
                "access_id": {
                    "type": "integer"
                },
                "has_access": {
                    "type": "boolean"
                }
            }
        },
        "be-dashboard-nba_internal_presentation_role_presenter.UpdateRoleAccessRequest": {
            "type": "object",
            "required": [
                "access_item"
            ],
            "properties": {
                "access_item": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/be-dashboard-nba_internal_presentation_role_presenter.UpdateRoleAccessItem"
                    }
                }
            }
        },
        "be-dashboard-nba_internal_presentation_role_presenter.UpdateRoleRequest": {
            "type": "object",
            "required": [
                "code",
                "name"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer {your token}\" into the field below.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}